# docs/frontend-backend/openapi-snapshot.yaml
openapi: 3.1.0
info:
  title: BigHead API
  version: 0.3.0
  description: 'Contrato canônico Sprint 3: matriz T01-T56 e implementação FastAPI.'
servers:
- url: http://localhost:8000
  description: Local
paths:
...
  /v1/approvals:
    get:
      tags:
      - approvals
      summary: Approvals
      operationId: approvals_v1_approvals_get
      security:
      - HTTPBearer: []
      parameters:
      - name: queue
        in: query
        required: false
        schema:
          enum:
          - pending
          - overdue
          - decided
          - all
          type: string
          default: pending
          title: Queue
      - name: risk
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Risk
      - name: dueBefore
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Duebefore
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 100
          title: Limit
      - name: x-organization-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Organization-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '403':
          $ref: '#/components/responses/Problem'
      x-bighead-screen: T20
      x-bighead-role: reviewer
      x-bighead-cache-event: inbox ordenada; `approvals.updated`
      x-bighead-critical-error: autoaprovacao bloqueada
  /v1/approvals/{approvalId}:
    get:
      tags:
      - approvals
      summary: Approval Detail
      operationId: approval_detail_v1_approvals__approvalId__get
      security:
      - HTTPBearer: []
      parameters:
      - name: approvalId
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Approvalid
      - name: x-organization-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Organization-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalDetailResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
      x-bighead-screen: T21
      x-bighead-role: reviewer
      x-bighead-cache-event: decisao imutavel, auditada; `approvals.decided`
      x-bighead-critical-error: autoaprovacao `403`; concorrencia `409`
  /v1/approvals/{approvalId}/scorecard:
    get:
      tags:
      - approvals
      summary: Scorecard
      operationId: scorecard_v1_approvals__approvalId__scorecard_get
      security:
      - HTTPBearer: []
      parameters:
      - name: approvalId
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Approvalid
      - name: x-organization-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Organization-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Scorecard V1 Approvals  Approvalid  Scorecard Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '403':
          $ref: '#/components/responses/Problem'
      x-bighead-screen: T22
      x-bighead-role: reviewer
      x-bighead-cache-event: sem cache longo
      x-bighead-critical-error: score reprovado sem motivo explicito
  /v1/policies/approvals:
    get:
      tags:
      - approvals
      summary: Get Policy
      operationId: get_policy_v1_policies_approvals_get
      security:
      - HTTPBearer: []
      parameters:
      - name: x-organization-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Organization-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalPolicyResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '409':
          $ref: '#/components/responses/Problem'
      x-bighead-screen: T23
      x-bighead-role: owner/admin
      x-bighead-cache-event: versao de politica; `policy.updated`
      x-bighead-critical-error: lacuna de risco na simulacao
    patch:
      tags:
      - approvals
      summary: Patch Policy
      operationId: patch_policy_v1_policies_approvals_patch
      security:
      - HTTPBearer: []
      parameters:
      - name: x-organization-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Organization-Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApprovalPolicyPatchRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalPolicyResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '409':
          $ref: '#/components/responses/Problem'
      x-bighead-screen: T23
      x-bighead-role: owner/admin
      x-bighead-cache-event: versao de politica; `policy.updated`
      x-bighead-critical-error: lacuna de risco na simulacao
  /v1/portal/items/{token}:
    get:
      tags:
      - portal
      summary: Portal Item
      operationId: portal_item_v1_portal_items__token__get
      parameters:
      - name: token
        in: path
        required: true
        schema:
          type: string
          title: Token
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Portal Item V1 Portal Items  Token  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '404':
          $ref: '#/components/responses/Problem'
        '410':
          $ref: '#/components/responses/Problem'
      x-bighead-screen: T24
      x-bighead-role: guest
      x-bighead-cache-event: token com uso; `portal.viewed`
      x-bighead-critical-error: token expirado, revogado ou usado
  /v1/agents:
    get:
      tags:
      - agents
      summary: Agents
      operationId: agents_v1_agents_get
      security:
      - HTTPBearer: []
      parameters:
      - name: x-organization-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Organization-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '403':
          $ref: '#/components/responses/Problem'
      x-bighead-screen: T25
      x-bighead-role: admin
      x-bighead-cache-event: filtro por status; `agents.updated`
      x-bighead-critical-error: agente arquivado ainda em uso
  /v1/agents/{agentId}:
    get:
      tags:
      - agents
      summary: Agent Detail
      operationId: agent_detail_v1_agents__agentId__get
      security:
      - HTTPBearer: []
      parameters:
      - name: agentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Agentid
      - name: x-organization-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Organization-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Agent Detail V1 Agents  Agentid  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
      x-bighead-screen: T26
      x-bighead-role: admin
      x-bighead-cache-event: analise de impacto; `agents.updated`
      x-bighead-critical-error: skill dependente quebrada
    patch:
      tags:
      - agents
      summary: Patch Agent
      operationId: patch_agent_v1_agents__agentId__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: agentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Agentid
      - name: x-organization-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Organization-Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentPatchRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Patch Agent V1 Agents  Agentid  Patch
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
      x-bighead-screen: T26
      x-bighead-role: admin
      x-bighead-cache-event: analise de impacto; `agents.updated`
      x-bighead-critical-error: skill dependente quebrada
  /v1/skills:
    get:
      tags:
      - skills
      summary: Skills
      operationId: skills_v1_skills_get
      security:
      - HTTPBearer: []
      parameters:
      - name: x-organization-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Organization-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '403':
          $ref: '#/components/responses/Problem'
      x-bighead-screen: T27
      x-bighead-role: admin
      x-bighead-cache-event: health resumido; `skills.updated`
      x-bighead-critical-error: skill de alto risco sem aprovacao
  /v1/skills/{skillId}/validate:
    post:
      tags:
      - skills
      summary: Validate Skill
      operationId: validate_skill_v1_skills__skillId__validate_post
      security:
      - HTTPBearer: []
      parameters:
      - name: skillId
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Skillid
      - name: x-organization-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Organization-Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SkillValidateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillValidateResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '408':
          $ref: '#/components/responses/Problem'
      x-bighead-screen: T28
      x-bighead-role: admin
      x-bighead-cache-event: execucao fake assinc; `skills.validation.completed`
      x-bighead-critical-error: timeout
  /v1/models:
    get:
      tags:
      - models
      summary: Models
      operationId: models_v1_models_get
      security:
      - HTTPBearer: []
      parameters:
      - name: x-organization-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Organization-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Models V1 Models Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '403':
          $ref: '#/components/responses/Problem'
      x-bighead-screen: T29
      x-bighead-role: admin
      x-bighead-cache-event: pricing atual; `models.updated`
      x-bighead-critical-error: preco sem vigencia
  /v1/prompts:
    get:
      tags:
      - prompts
      summary: Prompts
      operationId: prompts_v1_prompts_get
      security:
      - HTTPBearer: []
      parameters:
      - name: x-organization-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Organization-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '403':
          $ref: '#/components/responses/Problem'
      x-bighead-screen: T30
      x-bighead-role: admin
      x-bighead-cache-event: versoes paginadas; `prompts.updated`
      x-bighead-critical-error: rollback invalido
  /v1/workflows:
    get:
      tags:
      - workflows
      summary: Workflows
      operationId: workflows_v1_workflows_get
      security:
      - HTTPBearer: []
      parameters:
      - name: x-organization-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Organization-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '403':
          $ref: '#/components/responses/Problem'
      x-bighead-screen: T31
      x-bighead-role: admin/manager
      x-bighead-cache-event: lista por status; `workflows.updated`
      x-bighead-critical-error: workflow invalido publicado
  /v1/workflows/{workflowId}/validate:
    post:
      tags:
      - workflows
      summary: Validate Workflow Endpoint
      operationId: validate_workflow_endpoint_v1_workflows__workflowId__validate_post
      security:
      - HTTPBearer: []
      parameters:
      - name: workflowId
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Workflowid
      - name: x-organization-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Organization-Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowValidateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowValidateResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '409':
          $ref: '#/components/responses/Problem'
      x-bighead-screen: T32
      x-bighead-role: admin
      x-bighead-cache-event: simulacao async; `workflows.validation.completed`
      x-bighead-critical-error: ciclo indevido
  /v1/workflows/{workflowId}/versions:
    get:
      tags:
      - workflows
      summary: Workflow Versions
      operationId: workflow_versions_v1_workflows__workflowId__versions_get
      security:
      - HTTPBearer: []
      parameters:
      - name: workflowId
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Workflowid
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
            minimum: 1
          - type: 'null'
          title: Cursor
      - name: includeDiff
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Includediff
      - name: x-organization-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Organization-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Workflow Versions V1 Workflows  Workflowid  Versions Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '403':
          $ref: '#/components/responses/Problem'
      x-bighead-screen: T33
      x-bighead-role: admin
      x-bighead-cache-event: diff versionado; `workflowVersions.updated`
      x-bighead-critical-error: rollback com runs antigos incorreto
  /v1/playbooks/{playbookId}/instantiate:
    post:
      tags:
      - playbooks
      summary: Instantiate
      operationId: instantiate_v1_playbooks__playbookId__instantiate_post
      security:
      - HTTPBearer: []
      parameters:
      - name: playbookId
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Playbookid
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 200
          pattern: .*\S.*
          title: Idempotency-Key
      - name: x-organization-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Organization-Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlaybookInstantiateRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaybookInstantiateResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '403':
          $ref: '#/components/responses/Problem'
      x-bighead-screen: T34
      x-bighead-role: member
      x-bighead-cache-event: cria task/workflow; `playbooks.instantiated`
      x-bighead-critical-error: parametro ausente
...
  /v1/approvals/{approvalId}/decisions:
    get:
      tags:
      - approvals
      summary: Approval Decisions
      operationId: approval_decisions_v1_approvals__approvalId__decisions_get
      security:
      - HTTPBearer: []
      parameters:
      - name: approvalId
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Approvalid
      - name: x-organization-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Organization-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalDecisionHistoryResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/approvals/{approvalId}/decision:
    post:
      tags:
      - approvals
      summary: Decide
      operationId: decide_v1_approvals__approvalId__decision_post
      security:
      - HTTPBearer: []
      parameters:
      - name: approvalId
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Approvalid
      - name: x-organization-id
        in: header
        required: true
        schema:
          type: string
          format: uuid
          title: X-Organization-Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApprovalDecisionRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalDecisionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
  responses:
    Problem:
      description: Erro no formato RFC 9457
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
  schemas:
...
    ApprovalDetailResponse:
      properties:
        approval:
          additionalProperties: true
          type: object
          title: Approval
        task:
          additionalProperties: true
          type: object
          title: Task
        requester:
          additionalProperties: true
          type: object
          title: Requester
        assignee:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Assignee
        artifact:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Artifact
        evidence:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Evidence
        impact:
          additionalProperties: true
          type: object
          title: Impact
        availableActions:
          items:
            type: string
          type: array
          title: Availableactions
        decisionBlockedReason:
          anyOf:
          - type: string
          - type: 'null'
          title: Decisionblockedreason
      type: object
      required:
      - approval
      - task
      - requester
      - impact
      title: ApprovalDetailResponse
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    Page:
      properties:
        items:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Items
        counters:
          additionalProperties:
            type: integer
          type: object
          title: Counters
        nextCursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Nextcursor
      type: object
      required:
      - items
      title: Page
...